requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas>=2.1.4       # Compatible con Python 3.13
numpy>=1.26.0       # Compatible con Python 3.13
lightgbm==3.3.5
requests==2.31.0
python-dateutil==2.8.2
gunicorn==21.2.0
python-dotenv==1.0.0
